refactor: split docs extractor and markdown tests
