+++ crates/ox_content_docs/src/extractor.rs
@@ -1618,1070 +1618,4 @@ impl<'a> DocVisitor<'a> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn extract_overloaded_function_marks_only_implementation_has_body() {
-        let source = r"
-/**
- * Define a plugin with extension.
- */
-export function plugin<E>(options: WithExt): WithExtResult;
-/**
- * Define a plugin without extension.
- */
-export function plugin(options: WithoutExt): WithoutExtResult;
-/**
- * Define a plugin.
- */
-export function plugin(options: any = {}): any {
-    return options;
-}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "plugin.ts", SourceType::ts()).unwrap();
-        let plugins = items.iter().filter(|item| item.name == "plugin").collect::<Vec<_>>();
-
-        assert_eq!(plugins.len(), 3);
-        // Overload signatures carry no body; only the implementation does.
-        assert!(!plugins[0].has_body);
-        assert!(!plugins[1].has_body);
-        assert!(plugins[2].has_body);
-    }
-
-    #[test]
-    fn test_extract_function() {
-        let source = r"
-/**
- * Adds two numbers together.
- * @param a - The first number
- * @param b - The second number
- * @returns The sum of a and b
- */
-export function add(a: number, b: number): number {
-    return a + b;
-}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "test.ts", SourceType::ts()).unwrap();
-
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].name, "add");
-        assert_eq!(items[0].kind, DocItemKind::Function);
-        assert!(items[0].exported);
-        assert!(items[0].doc.as_ref().unwrap().contains("Adds two numbers"));
-        assert_eq!(items[0].params.len(), 2);
-    }
-
-    #[test]
-    fn function_object_literal_parameter_preserves_type_and_members() {
-        let source = r"
-/**
- * Define a plugin.
- *
- * @param options - Plugin options.
- * @param options.id - Plugin id.
- * @param options.name - Plugin display name.
- * @param options.setup - Setup hook.
- */
-export function plugin<Id, Deps, PluginExt, MergedExtensions>(options: {
-    id: Id;
-    name?: string;
-    dependencies?: Deps;
-    setup?: (
-        ctx: Readonly<
-            PluginContext<MergedExtensions>
-        >
-    ) => Awaitable<void>;
-    extension: PluginExt;
-    onExtension?: OnPluginExtension<MergedExtensions>;
-}): PluginWithExtension<PluginExt>;
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "plugin.ts", SourceType::ts()).unwrap();
-        let plugin = items.iter().find(|item| item.name == "plugin").unwrap();
-
-        assert_eq!(plugin.params.len(), 7);
-        assert_eq!(plugin.params[0].name, "options");
-        let parent_type = plugin.params[0].type_annotation.as_deref().unwrap();
-        assert_ne!(parent_type, "{ ... }");
-        assert!(parent_type.contains("id: Id"));
-        assert!(parent_type.contains("name?: string"));
-        assert!(parent_type.contains(
-            "setup?: (ctx: Readonly<PluginContext<MergedExtensions>>) => Awaitable<void>"
-        ));
-        assert_eq!(plugin.params[0].description.as_deref(), Some("Plugin options."));
-
-        let names = plugin.params.iter().map(|param| param.name.as_str()).collect::<Vec<_>>();
-        assert_eq!(
-            names,
-            [
-                "options",
-                "options.id",
-                "options.name?",
-                "options.dependencies?",
-                "options.setup?",
-                "options.extension",
-                "options.onExtension?",
-            ]
-        );
-        assert_eq!(plugin.params[1].type_annotation.as_deref(), Some("Id"));
-        assert_eq!(plugin.params[1].description.as_deref(), Some("Plugin id."));
-        assert_eq!(plugin.params[2].description.as_deref(), Some("Plugin display name."));
-        assert_eq!(
-            plugin.params[4].type_annotation.as_deref(),
-            Some("(ctx: Readonly<PluginContext<MergedExtensions>>) => Awaitable<void>")
-        );
-        assert_eq!(plugin.params[4].description.as_deref(), Some("Setup hook."));
-        assert!(plugin.params[2].optional);
-        assert!(plugin.params[3].optional);
-        assert!(plugin.params[4].optional);
-        assert!(plugin.params[6].optional);
-    }
-
-    #[test]
-    fn destructured_parameter_uses_jsdoc_name_and_extracted_type() {
-        let source = r"
-/**
- * Resolve command line arguments.
- *
- * @param args - Argument schema.
- * @param tokens - Parsed tokens.
- * @param resolveArgs - Resolve options.
- */
-export declare function resolveArgs<A extends Args>(
-    args: A,
-    tokens: ArgToken[],
-    { shortGrouping, skipPositional, toKebab }?: ResolveArgs
-): void;
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "resolver.ts", SourceType::ts()).unwrap();
-        let resolve = items.iter().find(|item| item.name == "resolveArgs").unwrap();
-
-        assert_eq!(
-            resolve.params.iter().map(|param| param.name.as_str()).collect::<Vec<_>>(),
-            ["args", "tokens", "resolveArgs"]
-        );
-        assert_eq!(resolve.params[2].type_annotation.as_deref(), Some("ResolveArgs"));
-        assert!(resolve.params[2].optional);
-        assert_eq!(resolve.params[2].description.as_deref(), Some("Resolve options."));
-    }
-
-    #[test]
-    fn destructured_parameter_without_jsdoc_name_keeps_param_fallback() {
-        let source = r"
-/**
- * Run a command.
- */
-export declare function run({ cwd }: RunOptions): void;
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "runner.ts", SourceType::ts()).unwrap();
-        let run = items.iter().find(|item| item.name == "run").unwrap();
-
-        assert_eq!(run.params.len(), 1);
-        assert_eq!(run.params[0].name, "param");
-        assert_eq!(run.params[0].type_annotation.as_deref(), Some("RunOptions"));
-        assert_eq!(run.params[0].description, None);
-    }
-
-    #[test]
-    fn destructured_parameter_keeps_nested_param_tags_on_members() {
-        let source = r"
-/**
- * Run a command.
- *
- * @param options - Runtime options.
- * @param options.cwd - Working directory.
- */
-export declare function run({ cwd }: { cwd: string }): void;
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "runner.ts", SourceType::ts()).unwrap();
-        let run = items.iter().find(|item| item.name == "run").unwrap();
-
-        assert_eq!(
-            run.params.iter().map(|param| param.name.as_str()).collect::<Vec<_>>(),
-            ["options", "options.cwd"]
-        );
-        assert_eq!(run.params[0].type_annotation.as_deref(), Some("{ cwd: string }"));
-        assert_eq!(run.params[0].description.as_deref(), Some("Runtime options."));
-        assert_eq!(run.params[1].type_annotation.as_deref(), Some("string"));
-        assert_eq!(run.params[1].description.as_deref(), Some("Working directory."));
-    }
-
-    #[test]
-    fn function_return_type_literal_emits_return_members() {
-        let source = r"
-/**
- * Resolve arguments.
- * @returns Resolved args.
- */
-export function resolveArgs<A extends Args>(): {
-    values: ArgValues<A>;
-    positionals: string[];
-    error: AggregateError | undefined;
-} {
-    return {} as any;
-}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "resolver.ts", SourceType::ts()).unwrap();
-
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].return_type.as_deref(), Some("object"));
-        assert_eq!(
-            items[0].return_members.iter().map(|member| member.name.as_str()).collect::<Vec<_>>(),
-            ["values", "positionals", "error"]
-        );
-        assert_eq!(items[0].return_members[0].signature.as_deref(), Some("ArgValues<A>"));
-        assert_eq!(items[0].return_members[1].signature.as_deref(), Some("string[]"));
-        assert_eq!(
-            items[0].return_members[2].signature.as_deref(),
-            Some("AggregateError | undefined")
-        );
-    }
-
-    #[test]
-    fn index_signatures_are_extracted_from_members_and_return_literals() {
-        let source = r"
-/**
- * Value type.
- */
-export interface ArgSchema {}
-
-/**
- * Arguments.
- */
-export interface Args {
-    /** Argument schema by option name. */
-    readonly [option: string]: ArgSchema;
-}
-
-/**
- * Numeric arguments.
- */
-export type NumericArgs = {
-    [index: number]: ArgSchema;
-};
-
-/**
- * Argument store.
- */
-export class Store {
-    [key: string]: ArgSchema;
-}
-
-/**
- * Makes arguments.
- */
-export function makeArgs(): {
-    [key: string]: ArgSchema;
-} {
-    return {} as any;
-}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "args.ts", SourceType::ts()).unwrap();
-        let args = items.iter().find(|item| item.name == "Args").unwrap();
-        let numeric_args = items.iter().find(|item| item.name == "NumericArgs").unwrap();
-        let store = items.iter().find(|item| item.name == "Store").unwrap();
-        let make_args = items.iter().find(|item| item.name == "makeArgs").unwrap();
-
-        let args_member = &args.children[0];
-        assert_eq!(args_member.kind, DocItemKind::IndexSignature);
-        assert_eq!(args_member.name, "[option: string]");
-        assert_eq!(args_member.signature.as_deref(), Some("readonly [option: string]: ArgSchema"));
-        assert_eq!(args_member.return_type.as_deref(), Some("ArgSchema"));
-        assert_eq!(args_member.params[0].name, "option");
-        assert_eq!(args_member.params[0].type_annotation.as_deref(), Some("string"));
-        assert!(args_member.readonly);
-
-        let numeric_member = &numeric_args.children[0];
-        assert_eq!(numeric_member.kind, DocItemKind::IndexSignature);
-        assert_eq!(numeric_member.name, "[index: number]");
-        assert_eq!(numeric_member.signature.as_deref(), Some("[index: number]: ArgSchema"));
-
-        let store_member = &store.children[0];
-        assert_eq!(store_member.kind, DocItemKind::IndexSignature);
-        assert_eq!(store_member.signature.as_deref(), Some("[key: string]: ArgSchema"));
-
-        let return_member = &make_args.return_members[0];
-        assert_eq!(return_member.kind, DocItemKind::IndexSignature);
-        assert_eq!(return_member.signature.as_deref(), Some("[key: string]: ArgSchema"));
-    }
-
-    #[test]
-    fn test_extract_interface() {
-        let source = r"
-/**
- * User interface.
- */
-export interface User {
-    /** User's name */
-    name: string;
-    /** User's age */
-    age: number;
-}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "test.ts", SourceType::ts()).unwrap();
-
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].name, "User");
-        assert_eq!(items[0].kind, DocItemKind::Interface);
-        assert_eq!(items[0].children.len(), 2);
-    }
-
-    #[test]
-    fn extracts_interface_extends_and_class_implements() {
-        let source = r"
-/**
- * Base runtime adapter.
- */
-export interface BaseAdapter {}
-
-/**
- * Runtime adapter.
- */
-export interface TranslationAdapter extends BaseAdapter {
-    /**
-     * Gets a locale resource.
-     * @param locale - Locale name.
-     * @returns The locale resource.
-     */
-    getResource(locale: string): Record<string, string> | undefined;
-}
-
-/**
- * Default runtime adapter.
- */
-export class DefaultTranslation implements TranslationAdapter {
-    /**
-     * Gets a locale resource.
-     * @param locale - Locale name.
-     * @returns The locale resource.
-     */
-    getResource(locale: string): Record<string, string> | undefined {
-        return undefined;
-    }
-}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "adapter.ts", SourceType::ts()).unwrap();
-        let adapter = items.iter().find(|item| item.name == "TranslationAdapter").unwrap();
-        let implementation = items.iter().find(|item| item.name == "DefaultTranslation").unwrap();
-
-        assert_eq!(adapter.kind, DocItemKind::Interface);
-        assert_eq!(adapter.extends, vec!["BaseAdapter"]);
-        assert_eq!(implementation.kind, DocItemKind::Class);
-        assert_eq!(implementation.implements, vec!["TranslationAdapter"]);
-        assert_eq!(
-            implementation.signature.as_deref(),
-            Some("export class DefaultTranslation implements TranslationAdapter")
-        );
-    }
-
-    #[test]
-    fn type_alias_object_literal_emits_property_children() {
-        let source = r"
-/**
- * Command options.
- */
-export type CommandOptions = {
-    name: string;
-    aliases?: string[];
-};
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "options.ts", SourceType::ts()).unwrap();
-
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].name, "CommandOptions");
-        assert_eq!(items[0].children.len(), 2);
-        assert_eq!(items[0].children[0].name, "name");
-        assert_eq!(items[0].children[0].kind, DocItemKind::Property);
-        assert_eq!(items[0].children[0].signature.as_deref(), Some("string"));
-        assert!(!items[0].children[0].optional);
-        assert_eq!(items[0].children[1].name, "aliases");
-        assert_eq!(items[0].children[1].signature.as_deref(), Some("string[]"));
-        assert!(items[0].children[1].optional);
-    }
-
-    #[test]
-    fn interface_property_type_literal_emits_property_members() {
-        let source = r"
-/**
- * Request options.
- */
-export interface RequestOptions {
-    /** HTTP options. */
-    http: {
-        /** Request timeout. */
-        timeout?: number;
-        /** Request headers. */
-        headers: Record<string, string>;
-    };
-}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "request.ts", SourceType::ts()).unwrap();
-        let http = &items[0].children[0];
-
-        assert_eq!(http.name, "http");
-        assert_eq!(http.kind, DocItemKind::Property);
-        assert_eq!(http.children.len(), 2);
-        assert_eq!(http.children[0].name, "timeout");
-        assert_eq!(http.children[0].doc.as_deref(), Some("Request timeout."));
-        assert_eq!(http.children[0].signature.as_deref(), Some("number"));
-        assert!(http.children[0].optional);
-        assert_eq!(http.children[1].name, "headers");
-        assert_eq!(http.children[1].signature.as_deref(), Some("Record<string, string>"));
-    }
-
-    #[test]
-    fn type_alias_signature_omits_nested_property_jsdoc_comments() {
-        let source = r"
-/**
- * A combinator produced by combinator factory functions.
- */
-export type Combinator<T> = {
-    /**
-     * The parse function that converts a string to the desired type.
-     *
-     * @param value - The input string value.
-     * @returns The parsed value of type T.
-     */
-    parse: (value: string) => T;
-};
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "combinators.ts", SourceType::ts()).unwrap();
-        let combinator = items.iter().find(|item| item.name == "Combinator").unwrap();
-        let signature = combinator.signature.as_deref().unwrap();
-
-        assert!(!signature.contains("/**"));
-        assert!(!signature.contains("@param"));
-        assert!(!signature.contains("@returns"));
-        assert_eq!(signature, "export type Combinator<T> = { parse: (value: string) => T }");
-
-        let parse = &combinator.children[0];
-        assert_eq!(
-            parse.doc.as_deref(),
-            Some("The parse function that converts a string to the desired type.")
-        );
-        assert_eq!(parse.signature.as_deref(), Some("(value: string) => T"));
-        assert_eq!(parse.params[0].name, "value");
-        assert_eq!(parse.params[0].description.as_deref(), Some("The input string value."));
-        assert_eq!(parse.return_type.as_deref(), Some("T"));
-        let returns_tag = parse.tags.iter().find(|tag| tag.tag == "returns").unwrap();
-        assert_eq!(returns_tag.description.as_deref(), Some("The parsed value of type T."));
-    }
-
-    #[test]
-    fn type_alias_object_literal_with_method_signature() {
-        let source = r"
-/**
- * Command options.
- */
-export type CommandOptions = {
-    /**
-     * Runs the command.
-     * @param ctx - Runtime context
-     */
-    run(ctx: Context): void;
-};
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "options.ts", SourceType::ts()).unwrap();
-        let signature = items[0].signature.as_deref().unwrap();
-        let member = &items[0].children[0];
-
-        assert!(!signature.contains("/**"));
-        assert!(!signature.contains("@param"));
-        assert_eq!(signature, "export type CommandOptions = { run(ctx: Context): void }");
-        assert_eq!(member.name, "run");
-        assert_eq!(member.kind, DocItemKind::Method);
-        assert_eq!(member.signature.as_deref(), Some("run(ctx: Context): void"));
-        assert_eq!(member.params.len(), 1);
-        assert_eq!(member.params[0].description.as_deref(), Some("Runtime context"));
-    }
-
-    #[test]
-    fn type_alias_intersection_extracts_object_literal_members() {
-        let source = r"
-/**
- * Command options.
- */
-export type CommandOptions = BaseOptions & {
-    /** Command name. */
-    name: string;
-};
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "options.ts", SourceType::ts()).unwrap();
-
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].name, "CommandOptions");
-        assert_eq!(items[0].children.len(), 1);
-        let signature = items[0].signature.as_deref().unwrap();
-        assert!(!signature.contains("/**"));
-        assert!(signature.contains("BaseOptions & { name: string }"));
-        assert_eq!(items[0].children[0].name, "name");
-        assert_eq!(items[0].children[0].signature.as_deref(), Some("string"));
-        assert!(items[0].signature.as_deref().unwrap().contains("BaseOptions &"));
-    }
-
-    #[test]
-    fn type_alias_intersection_resolves_callable_alias_and_members() {
-        let source = r"
-/**
- * Plugin function.
- */
-export type PluginFunction<G> = (ctx: Readonly<PluginContext<G>>) => Awaitable<void>;
-
-/**
- * Plugin.
- * @param ctx - Plugin context.
- * @returns Plugin setup result.
- */
-export type Plugin<E> = PluginFunction & {
-    id: string;
-    name?: string;
-};
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "plugin.ts", SourceType::ts()).unwrap();
-        let plugin = items.iter().find(|item| item.name == "Plugin").unwrap();
-
-        assert_eq!(plugin.params.len(), 1);
-        assert_eq!(plugin.params[0].name, "ctx");
-        assert_eq!(plugin.params[0].type_annotation.as_deref(), Some("Readonly<PluginContext<G>>"));
-        assert_eq!(plugin.params[0].description, None);
-        assert_eq!(plugin.return_type.as_deref(), Some("Awaitable<void>"));
-        assert_eq!(
-            plugin.children.iter().map(|child| child.name.as_str()).collect::<Vec<_>>(),
-            ["id", "name"]
-        );
-        assert_eq!(plugin.children[0].signature.as_deref(), Some("string"));
-        assert_eq!(plugin.children[1].signature.as_deref(), Some("string"));
-        assert!(plugin.children[1].optional);
-    }
-
-    #[test]
-    fn function_valued_interface_property_extracts_params_and_returns() {
-        let source = r"
-/**
- * Options for parsing.
- */
-export interface ArgSchema {
-    /**
-     * Parse a value.
-     */
-    parse?: (value: string) => any;
-}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "options.ts", SourceType::ts()).unwrap();
-        let schema = items.iter().find(|item| item.name == "ArgSchema").unwrap();
-        let parse = schema.children.iter().find(|member| member.name == "parse").unwrap();
-
-        assert_eq!(parse.kind, DocItemKind::Property);
-        assert_eq!(parse.signature.as_deref(), Some("(value: string) => any"));
-        assert_eq!(parse.params.len(), 1);
-        assert_eq!(parse.params[0].name, "value");
-        assert_eq!(parse.params[0].type_annotation.as_deref(), Some("string"));
-        assert_eq!(parse.return_type.as_deref(), Some("any"));
-    }
-
-    #[test]
-    fn function_valued_class_property_extracts_params_and_returns() {
-        let source = r"
-/**
- * Argument parser.
- */
-export class ArgParser {
-    /**
-     * Parse a value.
-     */
-    parse: (value: string) => any;
-}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "parser.ts", SourceType::ts()).unwrap();
-        let parser = items.iter().find(|item| item.name == "ArgParser").unwrap();
-        let parse = parser.children.iter().find(|member| member.name == "parse").unwrap();
-
-        assert_eq!(parse.kind, DocItemKind::Property);
-        assert_eq!(parse.signature.as_deref(), Some("(value: string) => any"));
-        assert_eq!(parse.params[0].name, "value");
-        assert_eq!(parse.params[0].type_annotation.as_deref(), Some("string"));
-        assert_eq!(parse.return_type.as_deref(), Some("any"));
-    }
-
-    #[test]
-    fn readonly_type_and_parenthesized_union_preserve_types() {
-        let source = r"
-/**
- * Command arguments.
- */
-export interface ArgSchema {
-    /**
-     * Parse a value.
-     */
-    choices?: string[] | readonly string[];
-}
-
-/**
- * Example rendering hooks.
- */
-export interface SubCommandable {
-    examples?: string | ((...args: any[]) => any);
-}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "schemas.ts", SourceType::ts()).unwrap();
-        let arg_schema = items.iter().find(|item| item.name == "ArgSchema").unwrap();
-        let sub = items.iter().find(|item| item.name == "SubCommandable").unwrap();
-
-        let choices = arg_schema.children.iter().find(|member| member.name == "choices").unwrap();
-        assert_eq!(choices.signature.as_deref(), Some("string[] | readonly string[]"));
-
-        let examples = sub.children.iter().find(|member| member.name == "examples").unwrap();
-        assert_eq!(examples.signature.as_deref(), Some("string | ((...args: any[]) => any)"));
-    }
-
-    #[test]
-    fn type_alias_function_extracts_params_and_returns() {
-        let source = r"
-/**
- * Run a command.
- */
-export type CommandRunner<G> = (ctx: Readonly<CommandContext<G>>) => Awaitable<string | void>;
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "runner.ts", SourceType::ts()).unwrap();
-        let alias = items.iter().find(|item| item.name == "CommandRunner").unwrap();
-
-        assert_eq!(alias.params.len(), 1);
-        assert_eq!(alias.params[0].name, "ctx");
-        assert_eq!(alias.params[0].type_annotation.as_deref(), Some("Readonly<CommandContext<G>>"));
-        assert_eq!(alias.return_type.as_deref(), Some("Awaitable<string | void>"));
-    }
-
-    #[test]
-    fn type_alias_function_with_multiple_parameters_extracts_all_params_and_return() {
-        let source = r"
-/**
- * Extend a command.
- */
-export type PluginExtension<T, G> = (ctx: CommandContextCore<G>, cmd: Command<G>) => Awaitable<T>;
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "plugin.ts", SourceType::ts()).unwrap();
-        let alias = items.iter().find(|item| item.name == "PluginExtension").unwrap();
-
-        assert_eq!(alias.params.len(), 2);
-        assert_eq!(alias.params[0].name, "ctx");
-        assert_eq!(alias.params[0].type_annotation.as_deref(), Some("CommandContextCore<G>"));
-        assert_eq!(alias.params[1].name, "cmd");
-        assert_eq!(alias.params[1].type_annotation.as_deref(), Some("Command<G>"));
-        assert_eq!(alias.return_type.as_deref(), Some("Awaitable<T>"));
-    }
-
-    #[test]
-    fn type_alias_function_with_jsdoc_params_but_no_returns_tag_extracts_return() {
-        let source = r"
-/**
- * Plugin extension hook.
- *
- * @param ctx - The command context.
- * @param cmd - The command.
- */
-export type OnPluginExtension<G> = (
-    ctx: Readonly<CommandContext<G>>,
-    cmd: Readonly<Command<G>>
-) => Awaitable<void>;
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "plugin.ts", SourceType::ts()).unwrap();
-        let alias = items.iter().find(|item| item.name == "OnPluginExtension").unwrap();
-
-        assert_eq!(alias.params.len(), 2);
-        assert_eq!(alias.params[0].name, "ctx");
-        assert_eq!(alias.params[0].type_annotation.as_deref(), Some("Readonly<CommandContext<G>>"));
-        assert_eq!(alias.params[0].description.as_deref(), Some("The command context."));
-        assert_eq!(alias.params[1].name, "cmd");
-        assert_eq!(alias.params[1].type_annotation.as_deref(), Some("Readonly<Command<G>>"));
-        assert_eq!(alias.params[1].description.as_deref(), Some("The command."));
-        assert_eq!(alias.return_type.as_deref(), Some("Awaitable<void>"));
-    }
-
-    #[test]
-    fn type_alias_function_with_function_param_and_return_extracts_nested_function_types() {
-        let source = r"
-/**
- * Decorate a runner.
- */
-export type CommandDecorator<G> = (
-    baseRunner: (ctx: Readonly<CommandContext<G>>) => Awaitable<string | void>
-) => (ctx: Readonly<CommandContext<G>>) => Awaitable<string | void>;
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "decorator.ts", SourceType::ts()).unwrap();
-        let alias = items.iter().find(|item| item.name == "CommandDecorator").unwrap();
-
-        assert_eq!(alias.params.len(), 1);
-        assert_eq!(alias.params[0].name, "baseRunner");
-        assert_eq!(
-            alias.params[0].type_annotation.as_deref(),
-            Some("(ctx: Readonly<CommandContext<G>>) => Awaitable<string | void>"),
-        );
-        assert_eq!(
-            alias.return_type.as_deref(),
-            Some("(ctx: Readonly<CommandContext<G>>) => Awaitable<string | void>"),
-        );
-    }
-
-    #[test]
-    fn type_alias_function_without_jsdoc_tags_still_extracts_metadata() {
-        let source = r"
-/**
- * Plugin function.
- */
-export type PluginFunction<G> = (ctx: Readonly<PluginContext<G>>) => Awaitable<void>;
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "plugin.ts", SourceType::ts()).unwrap();
-        let alias = items.iter().find(|item| item.name == "PluginFunction").unwrap();
-
-        assert_eq!(alias.params.len(), 1);
-        assert_eq!(alias.params[0].name, "ctx");
-        assert_eq!(alias.params[0].type_annotation.as_deref(), Some("Readonly<PluginContext<G>>"));
-        assert_eq!(alias.return_type.as_deref(), Some("Awaitable<void>"));
-    }
-
-    #[test]
-    fn test_extract_jsdoc_types_from_javascript() {
-        let source = r"
-/**
- * Creates a user-facing label.
- *
- * @param {string} value - The label source
- * @param {number} [maxLength=20] - Maximum length before truncation
- * @returns {string} Formatted label
- */
-export function label(value, maxLength = 20) {
-    return value.slice(0, maxLength);
-}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "test.js", SourceType::mjs()).unwrap();
-
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].name, "label");
-        assert_eq!(items[0].doc.as_deref(), Some("Creates a user-facing label."));
-        assert_eq!(items[0].return_type.as_deref(), Some("string"));
-        assert_eq!(items[0].params.len(), 2);
-        assert_eq!(items[0].params[0].type_annotation.as_deref(), Some("string"));
-        assert_eq!(items[0].params[0].description.as_deref(), Some("The label source"));
-        assert_eq!(items[0].params[1].type_annotation.as_deref(), Some("number"));
-        assert!(items[0].params[1].optional);
-        assert_eq!(items[0].params[1].default_value.as_deref(), Some("20"));
-        assert_eq!(
-            items[0].params[1].description.as_deref(),
-            Some("Maximum length before truncation")
-        );
-
-        let value_tag = items[0]
-            .tags
-            .iter()
-            .find(|tag| tag.tag == "param" && tag.name.as_deref() == Some("value"))
-            .unwrap();
-        assert_eq!(value_tag.type_annotation.as_deref(), Some("string"));
-        assert_eq!(value_tag.description.as_deref(), Some("The label source"));
-
-        let max_length_tag = items[0]
-            .tags
-            .iter()
-            .find(|tag| tag.tag == "param" && tag.name.as_deref() == Some("maxLength"))
-            .unwrap();
-        assert_eq!(max_length_tag.type_annotation.as_deref(), Some("number"));
-        assert_eq!(max_length_tag.optional, Some(true));
-        assert_eq!(max_length_tag.default_value.as_deref(), Some("20"));
-
-        let returns_tag = items[0].tags.iter().find(|tag| tag.tag == "returns").unwrap();
-        assert_eq!(returns_tag.type_annotation.as_deref(), Some("string"));
-        assert_eq!(returns_tag.description.as_deref(), Some("Formatted label"));
-    }
-
-    #[test]
-    fn test_extract_plain_top_level_variable() {
-        let source = r"
-/** Default placeholder when a command has no explicit name. */
-export const ANONYMOUS_COMMAND_NAME = '__anonymous__';
-
-/** Default retry count. */
-export let retries: number = 3;
-
-/** Creates labels. */
-export const label = (value: string): string => value;
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "constants.ts", SourceType::ts()).unwrap();
-
-        assert_eq!(items.len(), 3);
-        assert_eq!(items[0].name, "ANONYMOUS_COMMAND_NAME");
-        assert_eq!(items[0].kind, DocItemKind::Variable);
-        assert_eq!(
-            items[0].signature.as_deref(),
-            Some("export const ANONYMOUS_COMMAND_NAME = '__anonymous__'")
-        );
-        assert_eq!(items[1].name, "retries");
-        assert_eq!(items[1].kind, DocItemKind::Variable);
-        assert_eq!(items[1].signature.as_deref(), Some("export let retries: number"));
-        assert_eq!(items[2].name, "label");
-        assert_eq!(items[2].kind, DocItemKind::Function);
-    }
-
-    #[test]
-    fn test_undocumented_top_level_variable_is_skipped_by_default() {
-        let source = "export const ANONYMOUS_COMMAND_NAME = '(anonymous)';";
-
-        let items =
-            DocExtractor::new().extract_source(source, "constants.ts", SourceType::ts()).unwrap();
-
-        assert!(items.is_empty());
-    }
-
-    #[test]
-    fn test_extract_file_level_module_jsdoc() {
-        let source = r"
-/**
- * @module default
- *
- * Main entry point for the framework.
- */
-export { cli } from './core';
-
-/** Runs the CLI. */
-export function cli(): void {}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "src/index.ts", SourceType::ts()).unwrap();
-
-        assert_eq!(items.len(), 2);
-        assert_eq!(items[0].name, "default");
-        assert_eq!(items[0].kind, DocItemKind::Module);
-        assert_eq!(items[0].doc.as_deref(), Some("Main entry point for the framework."));
-        assert!(items[0].tags.iter().any(|tag| tag.tag == "module"));
-        assert_eq!(items[1].name, "cli");
-    }
-
-    #[test]
-    fn test_extract_function_type_parameters() {
-        let source = r"
-/** Make a thing. */
-export function make<G extends Base = Default, V>(value: V): G {
-  return value as unknown as G;
-}
-";
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "src/make.ts", SourceType::ts()).unwrap();
-        let func = items.iter().find(|item| item.name == "make").unwrap();
-
-        assert_eq!(func.type_parameters.len(), 2);
-        assert_eq!(func.type_parameters[0].name, "G");
-        assert_eq!(func.type_parameters[0].constraint.as_deref(), Some("Base"));
-        assert_eq!(func.type_parameters[0].default.as_deref(), Some("Default"));
-        assert_eq!(func.type_parameters[1].name, "V");
-        assert_eq!(func.type_parameters[1].constraint, None);
-        assert_eq!(func.type_parameters[1].default, None);
-    }
-
-    #[test]
-    fn test_extract_member_type_parameters() {
-        let source = r"
-/** Plugin context. */
-export interface PluginContext<G> {
-  /**
-   * Decorate the command.
-   * @typeParam L - Extension context.
-   */
-  decorateCommand<L extends Record<string, unknown> = DefaultExtensions>(
-    decorator: (value: L) => void
-  ): void;
-
-  /**
-   * Setup hook.
-   * @typeParam T - Hook value.
-   */
-  setup?: <T extends BaseHook = DefaultHook>(value: T) => Result;
-}
-";
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "src/context.ts", SourceType::ts()).unwrap();
-        let interface = items.iter().find(|item| item.name == "PluginContext").unwrap();
-
-        let method = interface.children.iter().find(|item| item.name == "decorateCommand").unwrap();
-        assert_eq!(method.type_parameters.len(), 1);
-        assert_eq!(method.type_parameters[0].name, "L");
-        assert_eq!(
-            method.type_parameters[0].constraint.as_deref(),
-            Some("Record<string, unknown>")
-        );
-        assert_eq!(method.type_parameters[0].default.as_deref(), Some("DefaultExtensions"));
-
-        let property = interface.children.iter().find(|item| item.name == "setup").unwrap();
-        assert_eq!(property.type_parameters.len(), 1);
-        assert_eq!(property.type_parameters[0].name, "T");
-        assert_eq!(property.type_parameters[0].constraint.as_deref(), Some("BaseHook"));
-        assert_eq!(property.type_parameters[0].default.as_deref(), Some("DefaultHook"));
-    }
-
-    #[test]
-    fn test_module_description_survives_trailing_author_comment() {
-        // Regression: `@module` block immediately followed by a second leading
-        // block comment (`@author`/`@license`). Both comments attach to the same
-        // first statement, so an `attached_to`-keyed lookup would surface the
-        // second comment and drop the `@module` description.
-        let source = r"
-/**
- * Module summary.
- * @module
- */
-
-/**
- * @author kazupon
- * @license MIT
- */
-export const z = 1;
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "src/context.ts", SourceType::ts()).unwrap();
-        let module = items.iter().find(|item| item.kind == DocItemKind::Module).unwrap();
-
-        assert_eq!(module.name, "context");
-        assert_eq!(module.doc.as_deref(), Some("Module summary."));
-        assert!(module.tags.iter().any(|tag| tag.tag == "module"));
-    }
-
-    #[test]
-    fn test_module_description_from_detached_comment_without_module_tag() {
-        // Gap 1: a leading file comment without `@module`, separated from the code
-        // by a blank line, should still be used as the module description
-        // (matching TypeDoc). The file stem becomes the module name.
-        let source = r"
-/**
- * The entry point for AI agent detection utility.
- *
- * @author kazupon
- * @license MIT
- */
-
-import { agentInfo } from 'std-env';
-
-/** A profile. */
-export function getAgentProfile(): void {}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "src/agent.ts", SourceType::ts()).unwrap();
-        let module = items.iter().find(|item| item.kind == DocItemKind::Module).unwrap();
-
-        assert_eq!(module.name, "agent");
-        assert_eq!(module.doc.as_deref(), Some("The entry point for AI agent detection utility."));
-        // The real declaration is still extracted with its own doc.
-        let func = items.iter().find(|item| item.name == "getAgentProfile").unwrap();
-        assert_eq!(func.doc.as_deref(), Some("A profile."));
-    }
-
-    #[test]
-    fn test_leading_comment_attached_to_declaration_is_not_a_module() {
-        // A doc comment that directly precedes the first declaration (no blank
-        // line, no module marker) documents that declaration, not the module.
-        let source = r"
-/** Documents foo. */
-export function foo(): void {}
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "src/foo.ts", SourceType::ts()).unwrap();
-
-        assert!(items.iter().all(|item| item.kind != DocItemKind::Module));
-        let func = items.iter().find(|item| item.name == "foo").unwrap();
-        assert_eq!(func.doc.as_deref(), Some("Documents foo."));
-    }
-
-    #[test]
-    fn test_module_jsdoc_name_falls_back_to_file_stem() {
-        let source = r"
-/**
- * @module
- */
-export { value } from './value';
-";
-
-        let extractor = DocExtractor::new();
-        let items = extractor.extract_source(source, "src/runtime.ts", SourceType::ts()).unwrap();
-
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].name, "runtime");
-        assert_eq!(items[0].kind, DocItemKind::Module);
-    }
-
-    #[test]
-    fn test_internal_items_are_excluded_by_default() {
-        let source = r"
-/** Public command. */
-export function publicCommand(): void {}
-
-/**
- * Internal helper.
- * @internal
- */
-export function internalHelper(): void {}
-";
-
-        let public_only =
-            DocExtractor::new().extract_source(source, "visibility.ts", SourceType::ts()).unwrap();
-        assert_eq!(public_only.len(), 1);
-        assert_eq!(public_only[0].name, "publicCommand");
-
-        let with_internal = DocExtractor::with_visibility(false, true)
-            .extract_source(source, "visibility.ts", SourceType::ts())
-            .unwrap();
-        assert_eq!(with_internal.len(), 2);
-        assert_eq!(with_internal[1].name, "internalHelper");
-    }
-}
+mod tests;
