+++ crates/ox_content_docs/src/extractor/tests.rs
@@ -1,0 +1,7 @@
+mod functions;
+mod jsdoc;
+mod member_shapes;
+mod modules;
+mod type_aliases;
+mod type_literals;
+mod type_members;
